requirements.txt
# Web应用安全测试工具依赖

# 基础依赖
requests>=2.31.0
beautifulsoup4>=4.12.2
lxml>=4.9.3

# 浏览器自动化
playwright>=1.40.0

# API和模型集成
openai>=1.6.0
python-dotenv>=1.0.0

# Web界面依赖
aiohttp>=3.8.6
websockets>=11.0.3
jinja2>=3.1.2

# 日志和解析
tqdm>=4.66.1
colorama>=0.4.6
tabulate>=0.9.0
rich>=13.6.0
pyfiglet>=0.8.0

# 安全功能
cryptography>=41.0.4
pyOpenSSL>=23.2.0
oauthlib>=3.2.2

# 测试工具
pytest>=7.4.3
pytest-cov>=4.1.0 